refactor:  refactor modules, separate eventListeners
