+++ src/js/calculator.js
@@ -1,34 +1,42 @@
 export default function calculator(rootComponent) {
   const $total = document.getElementById('total');
+
+  function render(total) {
+    $total.textContent = total;
+  }
+
+  initEventListeners(rootComponent, $total);
+
+  return {
+    render,
+  };
+}
+
+function initEventListeners(rootComponent, $total) {
   const $digits = document.querySelector('.digits');
   const $modifiers = document.querySelector('.modifiers');
   const $operations = document.querySelector('.operations');
 
-  function digitsClickEvent({ target }) {
+  function digitsClickEventListener({ target }) {
     const digit = target.textContent;
 
     rootComponent.dispatchEvent(new CustomEvent('digits', { detail: digit }));
   }
-  $digits.addEventListener('click', digitsClickEvent);
 
-  function operationsClickEvent({ target }) {
+  function operationsClickEventListener({ target }) {
     const operation = target.textContent;
 
     if (operation === '=') {
       rootComponent.dispatchEvent(new CustomEvent('calculate'));
-      const selected = $operations.querySelector('.selected');
-      if (selected) {
-        selected.classList.remove('selected');
-      }
+      clearOperations($operations);
       return;
     }
 
     target.classList.add('selected');
     rootComponent.dispatchEvent(new CustomEvent('op', { detail: operation }));
   }
-  $operations.addEventListener('click', operationsClickEvent);
 
-  function modifiersClickEvent({ target }) {
+  function modifiersClickEventListener({ target }) {
     const action = target.dataset.action;
 
     if (action === 'ac') {
@@ -36,21 +44,20 @@ export default function calculator(rootComponent) {
       rootComponent.dispatchEvent(new CustomEvent('ac'));
     }
   }
-  $modifiers.addEventListener('click', modifiersClickEvent);
 
   function clear() {
     $total.textValue = '0';
-    const selected = $operations.querySelector('selected');
+    clearOperations($operations);
+  }
+
+  function clearOperations($operations) {
+    const selected = $operations.querySelector('.selected');
     if (selected) {
       selected.classList.remove('selected');
     }
   }
 
-  function render(total) {
-    $total.textContent = total;
-  }
-
-  return {
-    render,
-  };
+  $digits.addEventListener('click', digitsClickEventListener);
+  $operations.addEventListener('click', operationsClickEventListener);
+  $modifiers.addEventListener('click', modifiersClickEventListener);
 }
